refactor: Move OffsetRecord to the offset-manager-api
- OffsetRecord is used by OffsetManager and SourceConnector. Not belong to the specific module.

- LogMetadata has different role against OffsetRecord. It consists of multiple metadata about the Log origin, offset, etc
