+++ offset-manager-api/src/main/java/offsetmanager/domain/OffsetStatus.java
@@ -7,7 +7,7 @@ import lombok.RequiredArgsConstructor;
 @Getter
 public enum OffsetStatus {
   INITIAL(0L),
-  COMPLETE(-1L);
+  COMPLETED(-1L);
 
   private final long value;
 }
