+++ offset-manager-remote/src/main/java/repository/RemoteOffsetManager.java
@@ -2,7 +2,7 @@ package repository;
 
 import jakarta.annotation.PreDestroy;
 import lombok.extern.slf4j.Slf4j;
-import offsetmanager.domain.OffsetManager;
+import offsetmanager.manager.OffsetManager;
 import org.apache.kafka.clients.consumer.Consumer;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
