+++ s3-source-connector/src/main/java/sourceconnector/domain/offset/LocalFileOffsetRecord.java
@@ -1,7 +1,9 @@
 package sourceconnector.domain.offset;
 
+import offsetmanager.domain.OffsetRecord;
+
 public record LocalFileOffsetRecord(
   String key,
   long offset
-) implements OffsetRecord{
+) implements OffsetRecord {
 }
