--- s3-source-connector/src/main/java/sourceconnector/domain/offset/OffsetRecord.java
@@ -1,6 +1,0 @@
-package sourceconnector.domain.offset;
-
-public interface OffsetRecord {
-  String key();
-  long offset();
-}
