+++ s3-source-connector/src/main/java/sourceconnector/domain/offset/S3OffsetRecord.java
@@ -1,7 +1,9 @@
 package sourceconnector.domain.offset;
 
+import offsetmanager.domain.OffsetRecord;
+
 public record S3OffsetRecord(
   String key,
   long offset
-) implements OffsetRecord{
+) implements OffsetRecord {
 }
