+++ s3-source-connector/src/main/java/sourceconnector/repository/OffsetRecordRepository.java
@@ -1,6 +1,6 @@
 package sourceconnector.repository;
 
-import sourceconnector.domain.offset.OffsetRecord;
+import offsetmanager.domain.OffsetRecord;
 
 import java.util.concurrent.ExecutionException;
