+++ s3-source-connector/src/main/java/sourceconnector/service/producer/BatchProducer.java
@@ -1,7 +1,8 @@
 package sourceconnector.service.producer;
 
+import offsetmanager.domain.OffsetRecord;
 import sourceconnector.domain.batch.MessageBatch;
-import sourceconnector.domain.offset.OffsetRecord;
+
 
 public interface BatchProducer<T> {
   void sendBatch(OffsetRecord offsetRecord, MessageBatch<T> messageBatch);
