+++ s3-source-connector/src/test/java/sourceconnector/repository/S3OffsetRecordRepositoryTest.java
@@ -14,7 +14,7 @@ import org.apache.kafka.common.config.TopicConfig;
 import org.apache.kafka.common.serialization.*;
 import org.junit.jupiter.api.*;
 import org.springframework.kafka.config.TopicBuilder;
-import sourceconnector.domain.offset.OffsetRecord;
+import offsetmanager.domain.OffsetRecord;
 
 import java.time.Duration;
 import java.util.Collections;
